refactor: Mid-refactor save point.
