+++ src/Concretize.hs
@@ -10,10 +10,12 @@ import Data.List (foldl')
 import Data.Maybe (fromMaybe)
 import Debug.Trace
 import Env (envIsExternal, findPoly, getTypeBinder, getValue, insert, insertX, lookupEverywhere, searchValue)
+import FindFunction
 import Info
 import InitialTypes
 import Managed
 import qualified Map
+import Memory (manageMemory)
 import Obj
 import Polymorphism
 import Reify
@@ -738,31 +740,6 @@ concretizeDefinition allowAmbiguity typeEnv globalEnv visitedDefinitions definit
         _ ->
           Left $ CannotConcretize definition
 
--- | Find ALL functions with a certain name, matching a type signature.
--- When the functionName argument denotes an interface, the name will match iff either:
---   1. The name of the binding matches functionName exactly OR
---   2. The name of the binding matches one of the names in the interface's implementation paths
--- For all other functions, the name must match exactly, and in all cases, the signature must match.
-allImplementations :: TypeEnv -> Env -> String -> Ty -> [(Env, Binder)]
-allImplementations typeEnv env functionName functionType =
-  (filter (predicate . xobjTy . binderXObj . snd) foundBindings)
-  where
-    predicate (Just t) =
-      --trace ("areUnifiable? " ++ show functionType ++ " == " ++ show t ++ " " ++ show (areUnifiable functionType t)) $
-      areUnifiable functionType t
-    predicate Nothing = error "allfunctionswithnameandsignature"
-    foundBindings = case getTypeBinder typeEnv functionName of
-      -- this function is an interface; lookup implementations
-      Right (Binder _ (XObj (Lst (XObj (Interface _ paths) _ _ : _)) _ _)) ->
-        case sequence $ map (\p -> searchValue env p) (paths ++ [(SymPath [] functionName)]) of
-          Right found -> found
-          Left _ ->
-            case findPoly env functionName functionType of
-              Right r -> [r]
-              Left _ -> (lookupEverywhere env functionName)
-      -- just a regular function; look for it
-      _ -> fromRight [] ((fmap (: []) (Env.getValue env functionName)) <> pure (lookupEverywhere env functionName))
-
 -- | Find all the dependencies of a polymorphic function with a name and a desired concrete type.
 depsOfPolymorphicFunction :: TypeEnv -> Env -> [SymPath] -> String -> Ty -> [XObj]
 depsOfPolymorphicFunction typeEnv env visitedDefinitions functionName functionType =
@@ -809,788 +786,6 @@ typesStrFunctionType typeEnv env memberType =
     then FuncTy [RefTy memberType (VarTy "q")] StringTy StaticLifetimeTy
     else FuncTy [memberType] StringTy StaticLifetimeTy
 
--- | The various results when trying to find a function using 'findFunctionForMember'.
-data FunctionFinderResult
-  = FunctionFound String
-  | FunctionNotFound String
-  | FunctionIgnored
-  deriving (Show)
-
--- | TODO: COMMENT THIS
-getConcretizedPath :: XObj -> Ty -> SymPath
-getConcretizedPath single functionType =
-  let Just t' = xobjTy single
-      (SymPath pathStrings name) = getPath single
-      suffix = polymorphicSuffix t' functionType
-   in SymPath pathStrings (name ++ suffix)
-
--- | Used for finding functions like 'delete' or 'copy' for members of a Deftype (or Array).
-findFunctionForMember :: TypeEnv -> Env -> String -> Ty -> (String, Ty) -> FunctionFinderResult
-findFunctionForMember typeEnv env functionName functionType (memberName, memberType)
-  | isManaged typeEnv env memberType =
-    case allImplementations typeEnv env functionName functionType of
-      [] ->
-        FunctionNotFound
-          ( "Can't find any '" ++ functionName ++ "' function for member '"
-              ++ memberName
-              ++ "' of type "
-              ++ show functionType
-          )
-      [(_, Binder _ single)] ->
-        let concretizedPath = getConcretizedPath single functionType
-         in FunctionFound (pathToC concretizedPath)
-      _ ->
-        FunctionNotFound
-          ( "Can't find a single '" ++ functionName ++ "' function for member '"
-              ++ memberName
-              ++ "' of type "
-              ++ show functionType
-          )
-  | otherwise = FunctionIgnored
-
--- | TODO: should this be the default and 'findFunctionForMember' be the specific one
-findFunctionForMemberIncludePrimitives :: TypeEnv -> Env -> String -> Ty -> (String, Ty) -> FunctionFinderResult
-findFunctionForMemberIncludePrimitives typeEnv env functionName functionType (memberName, _) =
-  case allImplementations typeEnv env functionName functionType of
-    [] ->
-      FunctionNotFound
-        ( "Can't find any '" ++ functionName ++ "' function for member '"
-            ++ memberName
-            ++ "' of type "
-            ++ show functionType
-        )
-    [(_, Binder _ single)] ->
-      let concretizedPath = getConcretizedPath single functionType
-       in FunctionFound (pathToC concretizedPath)
-    _ ->
-      FunctionNotFound
-        ( "Can't find a single '" ++ functionName ++ "' function for member '"
-            ++ memberName
-            ++ "' of type "
-            ++ show functionType
-        )
-
--- | Manage memory needs access to the concretizer
--- | (and the concretizer needs to manage memory)
--- | so they are put into the same module.
-
--- | Assign a set of Deleters to the 'infoDelete' field on Info.
-setDeletersOnInfo :: Maybe Info -> Set.Set Deleter -> Maybe Info
-setDeletersOnInfo i deleters = fmap (\i' -> i' {infoDelete = deleters}) i
-
-addDeletersToInfo :: Maybe Info -> Set.Set Deleter -> Maybe Info
-addDeletersToInfo i deleters =
-  fmap (\i' -> i' {infoDelete = Set.union (infoDelete i') deleters}) i
-
--- | Helper function for setting the deleters for an XObj.
-del :: XObj -> Set.Set Deleter -> XObj
-del xobj deleters = xobj {xobjInfo = setDeletersOnInfo (xobjInfo xobj) deleters}
-
--- | Differentiate between lifetimes depending on variables in a lexical scope and depending on something outside the function
-data LifetimeMode
-  = LifetimeInsideFunction String
-  | LifetimeOutsideFunction
-  deriving (Show)
-
--- | To keep track of the deleters when recursively walking the form.
-data MemState = MemState
-  { memStateDeleters :: Set.Set Deleter,
-    memStateDeps :: [XObj],
-    memStateLifetimes :: Map.Map String LifetimeMode
-  }
-  deriving (Show)
-
-prettyLifetimeMappings :: Map.Map String LifetimeMode -> String
-prettyLifetimeMappings mappings =
-  joinLines (map prettyMapping (Map.toList mappings))
-  where
-    prettyMapping (key, value) = "  " ++ key ++ " => " ++ show value
-
--- | Find out what deleters are needed and where in an XObj.
--- | Deleters will be added to the info field on XObj so that
--- | the code emitter can access them and insert calls to destructors.
-manageMemory :: TypeEnv -> Env -> XObj -> Either TypeError (XObj, [XObj])
-manageMemory typeEnv globalEnv root =
-  let (finalObj, finalState) = runState (visit root) (MemState (Set.fromList []) [] Map.empty)
-      deleteThese = memStateDeleters finalState
-      deps = memStateDeps finalState
-   in -- (trace ("Delete these: " ++ joinWithComma (map show (Set.toList deleteThese)))) $
-      case finalObj of
-        Left err -> Left err
-        Right ok ->
-          let newInfo = fmap (\i -> i {infoDelete = deleteThese}) (xobjInfo ok)
-           in -- This final check of lifetimes works on the lifetimes mappings after analyzing the function form, and
-              --  after all the local variables in it have been deleted. This is needed for values that are created
-              --  directly in body position, e.g. (defn f [] &[1 2 3])
-              case evalState (checkThatRefTargetIsAlive ok) (MemState (Set.fromList []) [] (memStateLifetimes finalState)) of
-                Left err -> Left err
-                Right _ -> Right (ok {xobjInfo = newInfo}, deps)
-  where
-    visit :: XObj -> State MemState (Either TypeError XObj)
-    visit xobj =
-      do
-        r <- case xobjObj xobj of
-          Lst _ -> visitList xobj
-          Arr _ -> visitArray xobj
-          StaticArr _ -> visitStaticArray xobj
-          Str _ -> do
-            manage xobj
-            addToLifetimesMappingsIfRef False xobj -- TODO: Should "internal = True" here? TODO: Possible to remove this one?
-            pure (Right xobj)
-          Pattern _ -> do
-            manage xobj
-            addToLifetimesMappingsIfRef False xobj -- TODO: Also possible to remove, *should* be superseeded by (***) below?
-            pure (Right xobj)
-          _ ->
-            pure (Right xobj)
-        case r of
-          Right ok -> do
-            MemState {} <- get
-            r' <- checkThatRefTargetIsAlive ok -- trace ("CHECKING " ++ pretty ok ++ " : " ++ showMaybeTy (ty xobj) ++ ", mappings: " ++ prettyLifetimeMappings m) $
-            addToLifetimesMappingsIfRef True ok -- (***)
-            pure r'
-          Left err -> pure (Left err)
-    visitArray :: XObj -> State MemState (Either TypeError XObj)
-    visitArray xobj@(XObj (Arr arr) _ _) =
-      do
-        mapM_ visit arr
-        results <- mapM unmanage arr
-        case sequence results of
-          Left e -> pure (Left e)
-          Right _ ->
-            do
-              _ <- manage xobj -- TODO: result is discarded here, is that OK?
-              pure (Right xobj)
-    visitArray _ = error "Must visit array."
-    visitStaticArray :: XObj -> State MemState (Either TypeError XObj)
-    visitStaticArray xobj@(XObj (StaticArr arr) _ _) =
-      do
-        mapM_ visit arr
-        results <- mapM unmanage arr
-        case sequence results of
-          Left e -> pure (Left e)
-          Right _ ->
-            -- We know that we want to add a deleter for the static array here
-            do
-              let var = varOfXObj xobj
-                  Just (RefTy t@(StructTy (ConcreteNameTy (SymPath [] "StaticArray")) [_]) _) = xobjTy xobj
-                  deleter = case nameOfPolymorphicFunction typeEnv globalEnv (FuncTy [t] UnitTy StaticLifetimeTy) "delete" of
-                    Just pathOfDeleteFunc ->
-                      ProperDeleter pathOfDeleteFunc (getDropFunc (xobjInfo xobj) t) var
-                    Nothing ->
-                      error ("No deleter found for Static Array : " ++ show t) --Just (FakeDeleter var)
-              MemState deleters deps lifetimes <- get
-              let newDeleters = Set.insert deleter deleters
-                  newDeps = deps ++ depsForDeleteFunc typeEnv globalEnv t
-                  newState = MemState newDeleters newDeps lifetimes
-              put newState --(trace (show newState) newState)
-              pure (Right xobj)
-    visitStaticArray _ = error "Must visit static array."
-    visitList :: XObj -> State MemState (Either TypeError XObj)
-    visitList xobj@(XObj (Lst lst) i t) =
-      case lst of
-        [defn@(XObj (Defn maybeCaptures) _ _), nameSymbol@(XObj (Sym _ _) _ _), args@(XObj (Arr argList) _ _), body] ->
-          let captures = maybe [] Set.toList maybeCaptures
-           in --case defnReturnType of
-              -- RefTy _ _ ->
-              --   pure (Left (FunctionsCantReturnRefTy xobj funcTy))
-              --  _ ->
-              do
-                mapM_ manage argList
-                -- Add the captured variables (if any, only happens in lifted lambdas) as fake deleters
-                -- TODO: Use another kind of Deleter for this case since it's pretty special?
-                mapM_
-                  ( ( \cap ->
-                        modify
-                          ( \memState ->
-                              memState {memStateDeleters = Set.insert (FakeDeleter cap) (memStateDeleters memState)}
-                          )
-                    )
-                      . getName
-                  )
-                  captures
-                mapM_ (addToLifetimesMappingsIfRef False) argList
-                mapM_ (addToLifetimesMappingsIfRef False) captures -- For captured variables inside of lifted lambdas
-                visitedBody <- visit body
-                result <- unmanage body
-                pure $
-                  case result of
-                    Left e -> Left e
-                    Right _ ->
-                      do
-                        okBody <- visitedBody
-                        pure (XObj (Lst [defn, nameSymbol, args, okBody]) i t)
-        -- Fn / λ (Lambda)
-        [fn@(XObj (Fn _ captures) _ _), args@(XObj (Arr _) _ _), body] ->
-          do
-            manage xobj -- manage inner lambdas but leave their bodies unvisited, they will be visited in the lifted version...
-            mapM_ unmanage captures
-            pure (Right (XObj (Lst [fn, args, body]) i t))
-        -- Def
-        [def@(XObj Def _ _), nameSymbol@(XObj (Sym _ _) _ _), expr] ->
-          do
-            visitedExpr <- visit expr
-            result <- unmanage expr
-            pure $
-              case result of
-                Left e -> Left e
-                Right () ->
-                  do
-                    okExpr <- visitedExpr
-                    pure (XObj (Lst [def, nameSymbol, okExpr]) i t)
-        -- Let
-        [letExpr@(XObj Let _ _), XObj (Arr bindings) bindi bindt, body] ->
-          do
-            MemState preDeleters _ _ <- get
-            visitedBindings <- mapM visitLetBinding (pairwise bindings)
-            visitedBody <- visit body
-            result <- unmanage body
-            case result of
-              Left e -> pure (Left e)
-              Right _ ->
-                do
-                  MemState postDeleters deps postLifetimes <- get
-                  let diff = postDeleters Set.\\ preDeleters
-                      newInfo = setDeletersOnInfo i diff
-                      survivors = postDeleters Set.\\ diff -- Same as just pre deleters, right?!
-                  put (MemState survivors deps postLifetimes)
-                  --trace ("LET Pre: " ++ show preDeleters ++ "\nPost: " ++ show postDeleters ++ "\nDiff: " ++ show diff ++ "\nSurvivors: " ++ show survivors)
-                  manage xobj
-                  pure $ do
-                    okBody <- visitedBody
-                    let finalBody = searchForInnerBreak diff okBody
-                    okBindings <- fmap (concatMap (\(n, x) -> [n, x])) (sequence visitedBindings)
-                    pure (XObj (Lst [letExpr, XObj (Arr okBindings) bindi bindt, finalBody]) newInfo t)
-        -- Set!
-        [setbangExpr@(XObj SetBang _ _), variable, value] ->
-          let varInfo = xobjInfo variable
-              correctVariableAndMode =
-                case variable of
-                  -- DISABLE FOR NOW: (XObj (Lst (XObj (Sym (SymPath _ "copy") _) _ _ : symObj@(XObj (Sym _ _) _ _) : _)) _ _) -> Right symObj
-                  symObj@(XObj (Sym _ mode) _ _) -> Right (symObj, mode)
-                  anythingElse -> Left (CannotSet anythingElse)
-           in case correctVariableAndMode of
-                Left err ->
-                  pure (Left err)
-                Right (okCorrectVariable, okMode) ->
-                  do
-                    MemState preDeleters _ _ <- get
-                    let ownsTheVarBefore = case createDeleter okCorrectVariable of
-                          Nothing -> Right ()
-                          Just d ->
-                            if Set.member d preDeleters || isLookupGlobal okMode
-                              then Right ()
-                              else Left (UsingUnownedValue variable)
-                    visitedValue <- visit value
-                    _ <- unmanage value -- The assigned value can't be used anymore
-                    MemState managed _ _ <- get
-                    -- Delete the value previously stored in the variable, if it's still alive
-                    let deleters = case createDeleter okCorrectVariable of
-                          Just d -> Set.fromList [d]
-                          Nothing -> Set.empty
-                        newVariable =
-                          case okMode of
-                            Symbol -> error "How to handle this?"
-                            LookupLocal _ ->
-                              if Set.size (Set.intersection managed deleters) == 1 -- The variable is still alive
-                                then variable {xobjInfo = setDeletersOnInfo varInfo deleters}
-                                else variable -- don't add the new info = no deleter
-                            LookupGlobal _ _ ->
-                              variable {xobjInfo = setDeletersOnInfo varInfo deleters}
-                            _ -> error "managememory set! 1"
-                    -- traceDeps = trace ("SET!-deleters for " ++ pretty xobj ++ " at " ++ prettyInfoFromXObj xobj ++ ":\n" ++
-                    --                    "unmanaged " ++ pretty value ++ "\n" ++
-                    --                    "managed: " ++ show managed ++ "\n" ++
-                    --                    "deleters: " ++ show deleters ++ "\n")
-
-                    case okMode of
-                      Symbol -> error "Should only be be a global/local lookup symbol."
-                      LookupLocal _ -> manage okCorrectVariable
-                      LookupGlobal _ _ -> pure ()
-                      _ -> error "managememory set! 2"
-                    pure $ case okMode of
-                      LookupLocal (Capture _) ->
-                        Left (CannotSetVariableFromLambda variable setbangExpr)
-                      _ ->
-                        do
-                          okValue <- visitedValue
-                          _ <- ownsTheVarBefore -- Force Either to fail
-                          pure (XObj (Lst [setbangExpr, newVariable, okValue]) i t)
-        [theExpr@(XObj The _ _), typeXObj, value] ->
-          do
-            visitedValue <- visit value
-            result <- transferOwnership value xobj
-            pure $ case result of
-              Left e -> Left e
-              Right _ -> do
-                okValue <- visitedValue
-                pure (XObj (Lst [theExpr, typeXObj, okValue]) i t)
-        [refExpr@(XObj Ref _ _), value] ->
-          do
-            visited <- visit value
-            case visited of
-              Left e -> pure (Left e)
-              Right visitedValue ->
-                do
-                  checkResult <- refCheck visitedValue
-                  case checkResult of
-                    Left e -> pure (Left e)
-                    Right () -> do
-                      let reffed = XObj (Lst [refExpr, visitedValue]) i t
-                      pure $ Right reffed
-        (XObj Deref _ _ : _) ->
-          error "Shouldn't end up here, deref only works when calling a function, i.e. ((deref f) 1 2 3)."
-        doExpr@(XObj Do _ _) : expressions ->
-          do
-            visitedExpressions <- mapM visit expressions
-            result <- transferOwnership (last expressions) xobj
-            pure $ case result of
-              Left e -> Left e
-              Right _ -> do
-                okExpressions <- sequence visitedExpressions
-                pure (XObj (Lst (doExpr : okExpressions)) i t)
-        [whileExpr@(XObj While _ _), expr, body] ->
-          do
-            MemState preDeleters _ _ <- get
-            visitedExpr <- visit expr
-            MemState afterExprDeleters _ _ <- get
-            visitedBody <- visit body
-            manage body
-            MemState postDeleters deps postLifetimes <- get
-            -- Visit an extra time to simulate repeated use
-            visitedExpr2 <- visit expr
-            visitedBody2 <- visit body
-            let diff = postDeleters \\ preDeleters
-            put (MemState (postDeleters \\ diff) deps postLifetimes) -- Same as just pre deleters, right?!
-            pure $ do
-              okExpr <- visitedExpr
-              okBody <- visitedBody
-              _ <- visitedExpr2 -- This evaluates the second visit so that it actually produces the error
-              _ <- visitedBody2 -- And this one too. Laziness FTW.
-              let newInfo = setDeletersOnInfo i diff
-                  -- Also need to set deleters ON the expression (for first run through the loop)
-                  XObj objExpr objInfo objTy = okExpr
-                  newExprInfo = setDeletersOnInfo objInfo (afterExprDeleters \\ preDeleters)
-                  newExpr = XObj objExpr newExprInfo objTy
-                  finalBody = searchForInnerBreak diff okBody
-              pure (XObj (Lst [whileExpr, newExpr, finalBody]) newInfo t)
-        [ifExpr@(XObj If _ _), expr, ifTrue, ifFalse] ->
-          do
-            visitedExpr <- visit expr
-            MemState preDeleters deps lifetimes <- get
-            let (visitedTrue, stillAliveTrue) =
-                  runState
-                    ( do
-                        v <- visit ifTrue
-                        result <- transferOwnership ifTrue xobj
-                        pure $ case result of
-                          Left e -> error (show e)
-                          Right () -> v
-                    )
-                    (MemState preDeleters deps lifetimes)
-                (visitedFalse, stillAliveFalse) =
-                  runState
-                    ( do
-                        v <- visit ifFalse
-                        result <- transferOwnership ifFalse xobj
-                        pure $ case result of
-                          Left e -> error (show e)
-                          Right () -> v
-                    )
-                    (MemState preDeleters deps lifetimes)
-            let deletedInTrue = preDeleters \\ memStateDeleters stillAliveTrue
-                deletedInFalse = preDeleters \\ memStateDeleters stillAliveFalse
-                deletedInBoth = Set.intersection deletedInTrue deletedInFalse
-                createdInTrue = memStateDeleters stillAliveTrue \\ preDeleters
-                createdInFalse = memStateDeleters stillAliveFalse \\ preDeleters
-                selfDeleter = case createDeleter xobj of
-                  Just ok -> Set.fromList [ok]
-                  Nothing -> Set.empty
-                createdAndDeletedInTrue = createdInTrue \\ selfDeleter
-                createdAndDeletedInFalse = createdInFalse \\ selfDeleter
-                delsTrue = Set.union (deletedInFalse \\ deletedInBoth) createdAndDeletedInTrue
-                delsFalse = Set.union (deletedInTrue \\ deletedInBoth) createdAndDeletedInFalse
-                stillAliveAfter = preDeleters \\ Set.union deletedInTrue deletedInFalse
-                depsAfter = memStateDeps stillAliveTrue ++ memStateDeps stillAliveFalse ++ deps -- Note: This merges all previous deps and the new ones, could be optimized..?!
-
-            -- traceDeps = trace ("IF-deleters for " ++ pretty xobj ++ " at " ++ prettyInfoFromXObj xobj ++ " " ++ identifierStr xobj ++ ":\n" ++
-            --                    "preDeleters: " ++ show preDeleters ++ "\n" ++
-            --                    "stillAliveTrue: " ++ show (memStateDeleters stillAliveTrue) ++ "\n" ++
-            --                    "stillAliveFalse: " ++ show (memStateDeleters stillAliveFalse) ++ "\n" ++
-            --                    "createdInTrue: " ++ show createdInTrue ++ "\n" ++
-            --                    "createdInFalse: " ++ show createdInFalse ++ "\n" ++
-            --                    "createdAndDeletedInTrue: " ++ show createdAndDeletedInTrue ++ "\n" ++
-            --                    "createdAndDeletedInFalse: " ++ show createdAndDeletedInFalse ++ "\n" ++
-            --                    "deletedInTrue: " ++ show deletedInTrue ++ "\n" ++
-            --                    "deletedInFalse: " ++ show deletedInFalse ++ "\n" ++
-            --                    "deletedInBoth: " ++ show deletedInBoth ++ "\n" ++
-            --                    "delsTrue: " ++ show delsTrue ++ "\n" ++
-            --                    "delsFalse: " ++ show delsFalse ++ "\n" ++
-            --                    "stillAlive: " ++ show stillAliveAfter ++ "\n" ++
-            --                    "depsAfter: " ++ show depsAfter ++ "\n"
-            --                   )
-
-            put (MemState stillAliveAfter depsAfter lifetimes)
-            manage xobj
-            pure $ do
-              okExpr <- visitedExpr
-              okTrue <- visitedTrue
-              okFalse <- visitedFalse
-              pure (XObj (Lst [ifExpr, okExpr, del okTrue delsTrue, del okFalse delsFalse]) i t)
-        matchExpr@(XObj (Match _) _ _) : expr : cases ->
-          -- General idea of how to figure out what to delete in a 'match' statement:
-          -- 1. Visit each case and investigate which variables are deleted in each one of the cases
-          -- 2. Variables deleted in at least one case has to be deleted in all, so make a union U of all such vars
-          --    but remove the ones that were not present before the 'match'
-          -- 3. In each case - take the intersection of U and the vars deleted in that case and add this result to its deleters
-          do
-            visitedExpr <- visit expr
-            case visitedExpr of
-              Left e -> pure (Left e)
-              Right okVisitedExpr ->
-                do
-                  _ <- unmanage okVisitedExpr
-                  MemState preDeleters deps lifetimes <- get
-                  vistedCasesAndDeps <- mapM visitMatchCase (pairwise cases)
-                  case sequence vistedCasesAndDeps of
-                    Left e -> pure (Left e)
-                    Right okCasesAndDeps ->
-                      let visitedCases = map fst okCasesAndDeps
-                          depsFromCases = concatMap snd okCasesAndDeps
-                          (finalXObj, postDeleters) = figureOutStuff okVisitedExpr visitedCases preDeleters
-                       in do
-                            put (MemState postDeleters (deps ++ depsFromCases) lifetimes)
-                            manage xobj
-                            pure (Right finalXObj)
-          where
-            figureOutStuff ::
-              XObj ->
-              [(Set.Set Deleter, (XObj, XObj))] ->
-              Set.Set Deleter ->
-              (XObj, Set.Set Deleter)
-            figureOutStuff okVisitedExpr visitedCasesWithDeleters preDeleters =
-              let postDeleters = map fst visitedCasesWithDeleters
-                  -- postDeletersUnion = unionOfSetsInList postDeleters
-                  postDeletersIntersection = intersectionOfSetsInList postDeleters
-                  deletersAfterTheMatch = Set.intersection preDeleters postDeletersIntersection
-                  -- The "postDeletersUnionPreExisting" are the vars that existed before the match but needs to
-                  -- be deleted after it has executed (because some branches delete them)
-                  -- postDeletersUnionPreExisting = Set.intersection postDeletersUnion preDeleters
-                  deletersForEachCase = map (\\ deletersAfterTheMatch) postDeleters
-                  -- These are the surviving vars after the 'match' expression:
-
-                  okVisitedCases = map snd visitedCasesWithDeleters
-                  okVisitedCasesWithAllDeleters =
-                    zipWith
-                      ( \(lhs, rhs) finalSetOfDeleters ->
-                          -- Putting the deleter info on the lhs,
-                          -- because the right one can collide with
-                          -- the other expressions, e.g. a 'let'
-                          let newLhsInfo = setDeletersOnInfo (xobjInfo lhs) finalSetOfDeleters
-                           in [lhs {xobjInfo = newLhsInfo}, rhs]
-                      )
-                      okVisitedCases
-                      deletersForEachCase
-               in -- trace ("post deleters: " ++ show postDeleters)
-                  -- trace ("\npost deleters union: " ++ show postDeletersUnion)
-                  -- trace ("\npost deleters intersection: " ++ show postDeletersIntersection)
-                  -- trace ("Post deleters union pre-existing: " ++ show postDeletersUnionPreExisting)
-                  -- trace ("Post deleters for each case: " ++ show postDeleters)
-                  ( XObj (Lst ([matchExpr, okVisitedExpr] ++ concat okVisitedCasesWithAllDeleters)) i t,
-                    deletersAfterTheMatch
-                  )
-        XObj (Lst [deref@(XObj Deref _ _), f]) xi xt : uargs ->
-          do
-            -- Do not visit f in this case, we don't want to manage it's memory since it is a ref!
-            visitedArgs <- sequence <$> mapM visitArg uargs
-            case visitedArgs of
-              Left err -> pure (Left err)
-              Right args ->
-                do
-                  unmanagedArgs <- sequence <$> mapM unmanageArg args
-                  manage xobj
-                  pure $ do
-                    okArgs <- unmanagedArgs
-                    Right (XObj (Lst (XObj (Lst [deref, f]) xi xt : okArgs)) i t)
-        f : uargs ->
-          do
-            visitedF <- visit f
-            visitedArgs <- sequence <$> mapM visitArg uargs
-            case visitedArgs of
-              Left err -> pure (Left err)
-              Right args -> do
-                unmanagedArgs <- sequence <$> mapM unmanageArg args
-                manage xobj
-                pure $ do
-                  okF <- visitedF
-                  okArgs <- unmanagedArgs
-                  Right (XObj (Lst (okF : okArgs)) i t)
-        [] -> pure (Right xobj)
-    visitList _ = error "Must visit list."
-    searchForInnerBreak :: Set.Set Deleter -> XObj -> XObj
-    searchForInnerBreak diff (XObj (Lst [(XObj Break i' t')]) xi xt) =
-      let ni = addDeletersToInfo i' diff
-       in XObj (Lst [(XObj Break ni t')]) xi xt
-    searchForInnerBreak _ x@(XObj (Lst ((XObj While _ _) : _)) _ _) = x
-    searchForInnerBreak diff (XObj (Lst elems) i' t') =
-      let newElems = map (searchForInnerBreak diff) elems
-       in XObj (Lst newElems) i' t'
-    searchForInnerBreak _ e = e
-    visitMatchCase :: (XObj, XObj) -> State MemState (Either TypeError ((Set.Set Deleter, (XObj, XObj)), [XObj]))
-    visitMatchCase (lhs@XObj {}, rhs@XObj {}) =
-      do
-        MemState preDeleters _ _ <- get
-        _ <- visitCaseLhs lhs
-        visitedRhs <- visit rhs
-        _ <- unmanage rhs
-        MemState postDeleters postDeps postLifetimes <- get
-        -- let diff = postDeleters \\ preDeleters
-        put (MemState preDeleters postDeps postLifetimes) -- Restore managed variables, TODO: Use a "local" state monad instead?
-        pure $ do
-          okVisitedRhs <- visitedRhs
-          -- trace ("\npre: " ++ show preDeleters ++
-          --        "\npost: " ++ show postDeleters ++
-          --        "\ndiff: " ++ show diff)
-          --   $
-          pure ((postDeleters, (lhs, okVisitedRhs)), postDeps)
-    visitCaseLhs :: XObj -> State MemState (Either TypeError [()])
-    visitCaseLhs (XObj (Lst vars) _ _) =
-      do
-        result <- mapM visitCaseLhs vars
-        let result' = sequence result
-        pure (fmap concat result')
-    visitCaseLhs xobj@(XObj (Sym (SymPath _ name) _) _ _)
-      | isVarName name = do
-        manage xobj
-        pure (Right [])
-      | otherwise = pure (Right [])
-    visitCaseLhs (XObj Ref _ _) =
-      pure (Right [])
-    visitCaseLhs x =
-      error ("Unhandled: " ++ show x)
-    addToLifetimesMappingsIfRef :: Bool -> XObj -> State MemState ()
-    addToLifetimesMappingsIfRef internal xobj =
-      case xobjTy xobj of
-        Just (RefTy _ (VarTy lt)) ->
-          do
-            m@(MemState _ _ lifetimes) <- get
-            case Map.lookup lt lifetimes of
-              Just _ ->
-                --trace ("\nThere is already a mapping for '" ++ pretty xobj ++ "' from the lifetime '" ++ lt ++ "' to " ++ show existing ++ ", won't add " ++ show (makeLifetimeMode xobj)) $
-                pure ()
-              Nothing ->
-                do
-                  let lifetimes' = Map.insert lt makeLifetimeMode lifetimes
-                  put $ --(trace $ "\nExtended lifetimes mappings for '" ++ pretty xobj ++ "' with " ++ show lt ++ " => " ++ show (makeLifetimeMode xobj) ++ " at " ++ prettyInfoFromXObj xobj ++ ":\n" ++ prettyLifetimeMappings lifetimes') $
-                    m {memStateLifetimes = lifetimes'}
-                  pure ()
-        Just _ ->
-          --trace ("Won't add to mappings! " ++ pretty xobj ++ " : " ++ show notThisType ++ " at " ++ prettyInfoFromXObj xobj) $
-          pure ()
-        _ ->
-          --trace ("No type on " ++ pretty xobj ++ " at " ++ prettyInfoFromXObj xobj) $
-          pure ()
-      where
-        makeLifetimeMode =
-          if internal
-            then LifetimeInsideFunction $
-              case xobj of
-                XObj (Lst [XObj Ref _ _, target]) _ _ -> varOfXObj target
-                _ -> varOfXObj xobj
-            else LifetimeOutsideFunction
-    checkThatRefTargetIsAlive :: XObj -> State MemState (Either TypeError XObj)
-    checkThatRefTargetIsAlive xobj =
-      -- TODO: Replace this whole thing with a function that collects all lifetime variables in a type.
-      case xobjTy xobj of
-        Just (RefTy _ (VarTy lt)) ->
-          performCheck lt
-        Just (FuncTy _ _ (VarTy lt)) ->
-          performCheck lt
-        -- HACK (not exhaustive):
-        Just (FuncTy _ (RefTy _ (VarTy lt)) _) ->
-          performCheck lt
-        _ ->
-          pure -- trace ("Won't check " ++ pretty xobj ++ " : " ++ show (ty xobj))
-            (Right xobj)
-      where
-        performCheck :: String -> State MemState (Either TypeError XObj)
-        performCheck lt =
-          do
-            MemState deleters _ lifetimeMappings <- get
-            case Map.lookup lt lifetimeMappings of
-              Just (LifetimeInsideFunction deleterName) ->
-                let matchingDeleters =
-                      Set.toList $
-                        Set.filter
-                          ( \case
-                              ProperDeleter {deleterVariable = dv} -> dv == deleterName
-                              FakeDeleter {deleterVariable = dv} -> dv == deleterName
-                              PrimDeleter {aliveVariable = dv} -> dv == deleterName
-                              RefDeleter {refVariable = dv} -> dv == deleterName
-                          )
-                          deleters
-                 in case matchingDeleters of
-                      [] ->
-                        --trace ("Can't use reference " ++ pretty xobj ++ " (with lifetime '" ++ lt ++ "', depending on " ++ show deleterName ++ ") at " ++ prettyInfoFromXObj xobj ++ ", it's not alive here:\n" ++ show xobj ++ "\nMappings: " ++ prettyLifetimeMappings lifetimeMappings ++ "\nAlive: " ++ show deleters ++ "\n") $
-                        --pure (Right xobj)
-                        pure (Left (UsingDeadReference xobj deleterName))
-                      _ ->
-                        --trace ("CAN use reference " ++ pretty xobj ++ " (with lifetime '" ++ lt ++ "', depending on " ++ show deleterName ++ ") at " ++ prettyInfoFromXObj xobj ++ ", it's not alive here:\n" ++ show xobj ++ "\nMappings: " ++ prettyLifetimeMappings lifetimeMappings ++ "\nAlive: " ++ show deleters ++ "\n") $
-                        pure (Right xobj)
-              Just LifetimeOutsideFunction ->
-                --trace ("Lifetime OUTSIDE function: " ++ pretty xobj ++ " at " ++ prettyInfoFromXObj xobj) $
-                pure (Right xobj)
-              Nothing ->
-                pure (Right xobj)
-    -- case xobj of
-    --   XObj (Sym _ (LookupLocal Capture)) _ _ ->
-    --     -- Ignore these for the moment! TODO: FIX!!!
-    --     pure (Right xobj)
-    --   _ ->
-    --     trace ("Failed to find lifetime key (when checking) '" ++ lt ++ "' for " ++ pretty xobj ++ " in mappings at " ++ prettyInfoFromXObj xobj) $
-    --     pure (Right xobj)
-
-    visitLetBinding :: (XObj, XObj) -> State MemState (Either TypeError (XObj, XObj))
-    visitLetBinding (name, expr) =
-      do
-        visitedExpr <- visit expr
-        addToLifetimesMappingsIfRef True expr
-        result <- transferOwnership expr name
-        pure $ case result of
-          Left e -> Left e
-          Right _ -> do
-            okExpr <- visitedExpr
-            pure (name, okExpr)
-    visitArg :: XObj -> State MemState (Either TypeError XObj)
-    visitArg xobj@(XObj _ _ (Just _)) =
-      do
-        afterVisit <- visit xobj
-        case afterVisit of
-          Right okAfterVisit -> do
-            addToLifetimesMappingsIfRef True okAfterVisit
-            pure (Right okAfterVisit)
-          Left err -> pure (Left err)
-    visitArg xobj@XObj {} =
-      visit xobj
-    unmanageArg :: XObj -> State MemState (Either TypeError XObj)
-    unmanageArg xobj@(XObj _ _ (Just t)) =
-      if isManaged typeEnv globalEnv t
-        then do
-          r <- unmanage xobj
-          pure $ case r of
-            Left err -> Left err
-            Right () -> Right xobj
-        else pure (Right xobj)
-    unmanageArg xobj@XObj {} =
-      pure (Right xobj)
-    getDropFunc i t = nameOfPolymorphicFunction typeEnv globalEnv (FuncTy [RefTy t (VarTy (makeTypeVariableNameFromInfo i))] UnitTy StaticLifetimeTy) "drop"
-    createDeleter xobj =
-      case xobjTy xobj of
-        Just (RefTy _ _) -> Just (RefDeleter (varOfXObj xobj))
-        Just t ->
-          let var = varOfXObj xobj
-           in if isManaged typeEnv globalEnv t
-                then case nameOfPolymorphicFunction typeEnv globalEnv (FuncTy [t] UnitTy StaticLifetimeTy) "delete" of
-                  Just pathOfDeleteFunc ->
-                    Just (ProperDeleter pathOfDeleteFunc (getDropFunc (xobjInfo xobj) t) var)
-                  Nothing ->
-                    --trace ("Found no delete function for " ++ var ++ " : " ++ (showMaybeTy (ty xobj)))
-                    Just (FakeDeleter var)
-                else Just (PrimDeleter var)
-        Nothing -> error ("No type, can't manage " ++ show xobj)
-    manage :: XObj -> State MemState ()
-    manage xobj =
-      if isSymbolThatCaptures xobj -- When visiting lifted lambdas, don't manage symbols that capture (they are owned by the environment).
-        then pure ()
-        else case createDeleter xobj of
-          Just deleter -> do
-            MemState deleters deps lifetimes <- get
-            let newDeleters = Set.insert deleter deleters
-                Just t = xobjTy xobj
-                newDeps = deps ++ depsForDeleteFunc typeEnv globalEnv t
-            put (MemState newDeleters newDeps lifetimes)
-          Nothing -> pure ()
-    deletersMatchingXObj :: XObj -> Set.Set Deleter -> [Deleter]
-    deletersMatchingXObj xobj deleters =
-      let var = varOfXObj xobj
-       in Set.toList $
-            Set.filter
-              ( \case
-                  ProperDeleter {deleterVariable = dv} -> dv == var
-                  FakeDeleter {deleterVariable = dv} -> dv == var
-                  PrimDeleter {aliveVariable = dv} -> dv == var
-                  RefDeleter {refVariable = dv} -> dv == var
-              )
-              deleters
-    isSymbolThatCaptures :: XObj -> Bool
-    isSymbolThatCaptures xobj =
-      case xobj of
-        XObj (Sym _ (LookupLocal (Capture _))) _ _ -> True
-        _ -> False
-    unmanage :: XObj -> State MemState (Either TypeError ())
-    unmanage xobj =
-      let Just t = xobjTy xobj
-       in if isManaged typeEnv globalEnv t && not (isGlobalFunc xobj)
-            then do
-              MemState deleters deps lifetimes <- get
-              case deletersMatchingXObj xobj deleters of
-                [] ->
-                  pure $
-                    if isSymbolThatCaptures xobj
-                      then Left (UsingCapturedValue xobj)
-                      else Left (UsingUnownedValue xobj)
-                [one] ->
-                  let newDeleters = Set.delete one deleters
-                   in do
-                        put (MemState newDeleters deps lifetimes)
-                        pure (Right ())
-                tooMany -> error ("Too many variables with the same name in set: " ++ show tooMany)
-            else pure (Right ())
-    refCheck :: XObj -> State MemState (Either TypeError ())
-    refCheck xobj =
-      let Just t = xobjTy xobj
-          isGlobalVariable = case xobj of
-            XObj (Sym _ (LookupGlobal _ _)) _ _ -> True
-            _ -> False
-       in if not isGlobalVariable && not (isGlobalFunc xobj) && isManaged typeEnv globalEnv t && not (isSymbolThatCaptures xobj) -- TODO: The 'isManaged typeEnv t' boolean check should be removed!
-            then do
-              MemState deleters _ _ <- get
-              pure $ case deletersMatchingXObj xobj deleters of
-                [] -> Left (GettingReferenceToUnownedValue xobj)
-                [_] -> pure ()
-                _ -> error $ "Too many variables with the same name in set (was looking for " ++ pretty xobj ++ " at " ++ prettyInfoFromXObj xobj ++ ")"
-            else pure (pure ())
-    transferOwnership :: XObj -> XObj -> State MemState (Either TypeError ())
-    transferOwnership from to =
-      do
-        result <- unmanage from
-        case result of
-          Left e -> pure (Left e)
-          Right _ -> do
-            manage to --(trace ("Transfered from " ++ getName from ++ " '" ++ varOfXObj from ++ "' to " ++ getName to ++ " '" ++ varOfXObj to ++ "'") to)
-            pure (Right ())
-
-varOfXObj :: XObj -> String
-varOfXObj xobj =
-  case xobj of
-    XObj (Sym path _) _ _ -> pathToC path
-    _ -> case xobjInfo xobj of
-      Just i -> freshVar i
-      Nothing -> error ("Missing info on " ++ show xobj)
-
-suffixTyVars :: String -> Ty -> Ty
-suffixTyVars suffix t =
-  case t of
-    VarTy key -> VarTy (key ++ suffix)
-    FuncTy argTys retTy ltTy -> FuncTy (map (suffixTyVars suffix) argTys) (suffixTyVars suffix retTy) (suffixTyVars suffix ltTy)
-    StructTy name tyArgs -> StructTy name (fmap (suffixTyVars suffix) tyArgs)
-    PointerTy x -> PointerTy (suffixTyVars suffix x)
-    RefTy x lt -> RefTy (suffixTyVars suffix x) (suffixTyVars suffix lt)
-    _ -> t
-
--- | The following functions will generate deleters and copy:ing methods for structs, they are shared with the Deftype module
-data AllocationMode = StackAlloc | HeapAlloc
-
 -- | The template for the 'delete' function of a concrete deftype.
 concreteDelete :: TypeEnv -> Env -> [(String, Ty)] -> Template
 concreteDelete typeEnv env members =
@@ -1633,21 +828,6 @@ concreteDeleteTakePtr typeEnv env members =
           (filter (isManaged typeEnv env) (map snd members))
     )
 
--- | Generate the C code for deleting a single member of the deftype.
--- | TODO: Should return an Either since this can fail!
-memberDeletionGeneral :: String -> TypeEnv -> Env -> (String, Ty) -> String
-memberDeletionGeneral separator typeEnv env (memberName, memberType) =
-  case findFunctionForMember typeEnv env "delete" (typesDeleterFunctionType memberType) (memberName, memberType) of
-    FunctionFound functionFullName -> "    " ++ functionFullName ++ "(p" ++ separator ++ memberName ++ ");"
-    FunctionNotFound msg -> error msg
-    FunctionIgnored -> "    /* Ignore non-managed member '" ++ memberName ++ "' : " ++ show memberType ++ " */"
-
-memberDeletion :: TypeEnv -> Env -> (String, Ty) -> String
-memberDeletion = memberDeletionGeneral "."
-
-memberRefDeletion :: TypeEnv -> Env -> (String, Ty) -> String
-memberRefDeletion = memberDeletionGeneral "Ref->"
-
 -- | The template for the 'copy' function of a concrete deftype.
 concreteCopy :: TypeEnv -> Env -> [(String, Ty)] -> Template
 concreteCopy typeEnv env memberPairs =
@@ -1661,27 +841,6 @@ concreteCopy typeEnv env memberPairs =
           (filter (isManaged typeEnv env) (map snd memberPairs))
     )
 
-tokensForCopy :: TypeEnv -> Env -> [(String, Ty)] -> [Token]
-tokensForCopy typeEnv env memberPairs =
-  toTemplate $
-    unlines
-      [ "$DECL {",
-        "    $p copy = *pRef;",
-        joinLines (map (memberCopy typeEnv env) memberPairs),
-        "    return copy;",
-        "}"
-      ]
-
--- | Generate the C code for copying the member of a deftype.
--- | TODO: Should return an Either since this can fail!
-memberCopy :: TypeEnv -> Env -> (String, Ty) -> String
-memberCopy typeEnv env (memberName, memberType) =
-  case findFunctionForMember typeEnv env "copy" (typesCopyFunctionType memberType) (memberName, memberType) of
-    FunctionFound functionFullName ->
-      "    copy." ++ memberName ++ " = " ++ functionFullName ++ "(&(pRef->" ++ memberName ++ "));"
-    FunctionNotFound msg -> error msg
-    FunctionIgnored -> "    /* Ignore non-managed member '" ++ memberName ++ "' : " ++ show memberType ++ " */"
-
 concreteCopyPtr :: TypeEnv -> Env -> [(String, Ty)] -> Template
 concreteCopyPtr typeEnv env memberPairs =
   Template
@@ -1693,23 +852,3 @@ concreteCopyPtr typeEnv env memberPairs =
           (depsOfPolymorphicFunction typeEnv env [] "copy" . typesCopyFunctionType)
           (filter (isManaged typeEnv env) (map snd memberPairs))
     )
-
-tokensForCopyPtr :: TypeEnv -> Env -> [(String, Ty)] -> [Token]
-tokensForCopyPtr typeEnv env memberPairs =
-  toTemplate $
-    unlines
-      [ "$DECL {",
-        "    $p* copy = CARP_MALLOC(sizeof(*pRef));",
-        "    *copy = *pRef;",
-        joinLines (map (memberCopyPtr typeEnv env) memberPairs),
-        "    return copy;",
-        "}"
-      ]
-
-memberCopyPtr :: TypeEnv -> Env -> (String, Ty) -> String
-memberCopyPtr typeEnv env (memberName, memberType) =
-  case findFunctionForMember typeEnv env "copy" (typesCopyFunctionType memberType) (memberName, memberType) of
-    FunctionFound functionFullName ->
-      "    copy->" ++ memberName ++ " = " ++ functionFullName ++ "(&(pRef->" ++ memberName ++ "));"
-    FunctionNotFound msg -> error msg
-    FunctionIgnored -> "    /* Ignore non-managed member '" ++ memberName ++ "' : " ++ show memberType ++ " */"
